requirements.txt
behave
pandas
openpyxl
requests
selenium
psycopg2-binary
allure-behave
jsonschema
python-dotenv
